Restructure of views + more folders
