+++ fitwsarah_frontend/src/README.txt
@@ -1,0 +1,19 @@
+1-The assets folder contains all of the project's static files, such as your logo, fonts, images, and favicons.
+
+2-Components folder contains a bit collection of UI codes such as buttons, forms, avatars, and so on.
+
+3-The views folder contains all your React application's web pages.
+
+4-Services folder contains code that allows you to interact with external API resources.
+
+5-The utils folder contains reusable function snippets for performing quick tasks like text truncation or down casing.
+
+6-Hooks folder contains codes and logic that can be reused across multiple components.
+
+7-The store folder houses your state management files, such as Redux, which are used to make certain functions and variables available throughout your application.
+
+8-The main component of your React application is the App.js file. This file connects all components and views.
+
+9-Index.js file is the React application's entry point. It is responsible for bootstrapping the React library and mounting it on the root element.
+
+10-index.css is the main and global CSS file for our application. Any writing style to this file will apply throughout the project.
