--- fitwsarah_frontend/src/views/home/home.html
@@ -1,6 +1,0 @@
-import '../../views/appointmentList';
-import '../../views/packageList';
-import '../../views/serviceList';
-
-
-
