--- fitwsarah_frontend/src/views/layout/layout.html
@@ -1,7 +1,0 @@
-import '../../views/footer';
-import '../../views/nav';
-
-
-
-
-
